adicionado barra de pesquisa

+++ vendas/views.py
@@ -7,6 +7,7 @@ from django.http import HttpResponse, JsonResponse
 import json
 from .models import *
 from .utils import dadosCarrinho, renderForm, temNone
+from django.db.models import Q
 
 from django.views.generic import (
     ListView,
@@ -205,6 +206,13 @@ def produtos(request):
             sweetify.error(
                 request, 'Porfavor cadastre seus dados antes de fazer um pedido')
             return redirect('perfil')
+    elif request.user.type == "LOJA":
+        try:
+            request.user.loja
+        except:
+            sweetify.error(
+                request, 'Porfavor cadastre seus dados antes de fazer um pedido')
+            return redirect('perfil')
     else:
         if revendedorPed == None:
             sweetify.error(request, 'Selecione um pedido para poder altera-lo')
@@ -260,11 +268,22 @@ def atualizarItem(request):
     action = data['action']
     if request.user.type == 'REVENDEDOR':
         revendedor = request.user.revendedor
+        pedido, created = Pedido.objects.get_or_create(
+        revendedor=revendedor, completo=False)
+    elif request.user.type == 'LOJA':
+        loja = request.user.loja
+        pedido, created = Pedido.objects.get_or_create(
+        loja=loja, completo=False)
     else:
         revendedor = revendedorPed
+        try:
+            pedido, created = Pedido.objects.get_or_create(
+            revendedor=revendedor, completo=False)
+        except:
+            pedido, created = Pedido.objects.get_or_create(
+            loja=revendedor, completo=False)
     produto = Produto.objects.get(id=idProduto)
-    pedido, created = Pedido.objects.get_or_create(
-        revendedor=revendedor, completo=False)
+    
 
     itemPedido, created = ItemPedido.objects.get_or_create(
         pedido=pedido, produto=produto)
@@ -287,17 +306,26 @@ def atualizarItem(request):
 
 
 def processarPedido(request):
-    cod_pedido = datetime.datetime.now().timestamp()
+    cod_pedido = str(datetime.datetime.now().timestamp())
     dados = json.loads(request.body)
 
     if request.user.is_authenticated:
         if request.user.type == 'REVENDEDOR':
             revendedor = request.user.revendedor
+            pedido, criado = Pedido.objects.get_or_create(
+                revendedor=revendedor, completo=False)
+        elif request.user.type == 'LOJA':
+            loja = request.user.loja
+            pedido, criado = Pedido.objects.get_or_create(
+                loja=loja, completo=False)
         else:
             revendedor = revendedorPed
-
-        pedido, criado = Pedido.objects.get_or_create(
-            revendedor=revendedor, completo=False)
+            try:
+                pedido, criado = Pedido.objects.get_or_create(
+                    revendedor=revendedor, completo=False)
+            except:
+                    pedido, criado = Pedido.objects.get_or_create(
+                    loja=revendedor, completo=False)
 
     total = float(dados['form']['total'].replace(',', '.'))
     pedido.cod_pedido = cod_pedido
@@ -323,14 +351,24 @@ def processarPedido(request):
 
 
 def mostrarPedidos(request):
-    try:
-        request.user.revendedor
-    except:
-        sweetify.error(
-            request, 'Porfavor cadastre seus dados antes de fazer um pedido')
-        return redirect('perfil')
-    pedidos = Pedido.objects.filter(revendedor=request.user.revendedor).exclude(
-        completo=False).order_by('-data')
+    if request.user.type == 'REVENDEDOR':
+        try:
+            request.user.revendedor
+        except:
+            sweetify.error(
+                request, 'Porfavor cadastre seus dados antes de fazer um pedido')
+            return redirect('perfil')
+        pedidos = Pedido.objects.filter(revendedor=request.user.revendedor).exclude(
+            completo=False).order_by('-data')
+    elif request.user.type == 'LOJA':
+        try:
+            request.user.loja
+        except:
+            sweetify.error(
+                request, 'Porfavor cadastre seus dados antes de fazer um pedido')
+            return redirect('perfil')
+        pedidos = Pedido.objects.filter(loja=request.user.loja).exclude(
+            completo=False).order_by('-data')
 
     return render(request, "vendas/meus_pedidos.html", {'pedidos': pedidos})
 
@@ -354,12 +392,10 @@ def atualizarPedido(request, pk):
         sweetify.success(
             request, 'Por favor altere o pedido a faça checkout novamente')
         return redirect('produtos')
-    else:
+    elif request.user.type == "LOJA":
         pedido = Pedido.objects.get(id=pk)
         pedido_ex = Pedido.objects.get(
-            revendedor=pedido.revendedor, completo=False)
-        global revendedorPed
-        revendedorPed = pedido.revendedor
+            loja=request.user.loja, completo=False)
         pedido_ex.delete()
         pedido.completo = False
         pedido.devolve_produtos()
@@ -367,6 +403,34 @@ def atualizarPedido(request, pk):
         sweetify.success(
             request, 'Por favor altere o pedido a faça checkout novamente')
         return redirect('produtos')
+    else:
+        global revendedorPed
+        try:
+            print('revendedor: ' + str(pedido.revendedor))
+            pedido = Pedido.objects.get(id=pk)
+            pedido_ex = Pedido.objects.get(
+                revendedor=pedido.revendedor, completo=False)
+            revendedorPed = pedido.revendedor
+            pedido_ex.delete()
+            pedido.completo = False
+            pedido.devolve_produtos()
+            pedido.save()
+            sweetify.success(
+                request, 'Por favor altere o pedido a faça checkout novamente')
+            return redirect('produtos')
+        except:
+            pedido = Pedido.objects.get(id=pk)
+            pedido_ex = Pedido.objects.get(
+                loja=pedido.loja, completo=False)
+            print('loja: ' + str(pedido.loja))
+            revendedorPed = pedido.loja
+            pedido_ex.delete()
+            pedido.completo = False
+            pedido.devolve_produtos()
+            pedido.save()
+            sweetify.success(
+                request, 'Por favor altere o pedido a faça checkout novamente')
+            return redirect('produtos')
 
 def deletarPedido(request, pk):
     pedido = Pedido.objects.get(id=pk)
@@ -436,3 +500,76 @@ def aprovarPedido(request, pk):
     pedido.save()
     sweetify.success(request, 'Pedido aprovado!')
     return redirect('pedidos')
+
+def enviarPedido(request, pk):
+    if pedido.status == 'APROVADO':
+        pedido = Pedido.objects.get(id=pk)
+        pedido.status = 'ENVIADO'
+        pedido.save()
+        sweetify.success(request, 'Pedido enviado!')
+        return redirect('pedidos')
+    else:
+        sweetify.error(request, 'O pedido ainda não foi aprovado!')
+
+def confirmarPedido(request, pk):
+    if pedido.status == 'ENVIADO':
+        pedido = Pedido.objects.get(id=pk)
+        pedido.status = 'FINALIZADO'
+        pedido.save()
+        sweetify.success(request, 'Pedido finalizado!')
+        return redirect('pedidos')
+    else:
+        sweetify.error(request, 'O pedido ainda não foi enviado!')
+
+class pesquisaUsuarios(ListView):
+    model = User
+
+    def get_queryset(self):
+        query = self.request.GET.get("q")
+        object_list = User.objects.filter(
+            Q(email__icontains=query)
+        )
+        return object_list
+
+class pesquisaPedidos(ListView):
+    model = Pedido
+
+
+    def get_queryset(self):
+        query = self.request.GET.get("q")
+        if self.request.user.type == 'REVENDEDOR':
+            object_list = Pedido.objects.filter(Q(revendedor=self.request.user.revendedor),
+                Q(cod_pedido__icontains=query) | Q(status__icontains=query)
+                | Q(metodo_de_pagamento__icontains=query) | Q(data__year=query)
+                | Q(data__month=query) | Q(data__day=query)
+            )
+        elif self.request.user.type == 'LOJA':
+            object_list = Pedido.objects.filter(Q(loja=self.request.user.loja),
+                Q(cod_pedido__icontains=query) | Q(status__icontains=query)
+                | Q(metodo_de_pagamento__icontains=query) | Q(data__year=query)
+                | Q(data__month=query) | Q(data__day=query)
+            )
+        else:
+            object_list = Pedido.objects.filter(
+                Q(cod_pedido__icontains=query) | Q(status__icontains=query)
+                | Q(metodo_de_pagamento__icontains=query) | Q(data__year=query)
+                | Q(data__month=query) | Q(data__day=query)
+            )
+        return object_list
+
+class pesquisaProdutos(ListView):
+    model = Produto
+
+    def get_context_data(self, **kwargs):
+        context = super(pesquisaProdutos, self).get_context_data(**kwargs)
+        user = self.request.user
+        context['user'] = user
+        return context
+
+    def get_queryset(self):
+        query = self.request.GET.get("q")
+        object_list = Produto.objects.filter(
+            Q(codigo__icontains=query) | Q(nome__icontains=query)
+        )
+
+        return object_list
